implemented trasition component for mobile, modified storybook
