+++ src/app/_components/Filter/Filter.tsx
@@ -1,5 +1,7 @@
-import React, { useEffect, useState } from "react";
-import { Dialog, Disclosure } from "@headlessui/react";
+"use client";
+
+import React, { Fragment, useEffect, useState } from "react";
+import { Dialog, Disclosure, Transition } from "@headlessui/react";
 import {
   ChevronDownIcon,
   AdjustmentsHorizontalIcon,
@@ -27,66 +29,11 @@ interface Filter {
   options: FilterOption[];
 }
 
-const subCategories = [
-  { name: "Clothes" },
-  { name: "Shoes" },
-  { name: "Accessories" },
-];
+interface FilterProps {
+  filters: Filter[];
+}
 
-const initialFilters = [
-  {
-    id: "color",
-    name: "Color",
-    options: [
-      { value: "white", label: "White", checked: false, color: "#FFFFFF" },
-      { value: "beige", label: "Beige", checked: false, color: "#F5F5DC" },
-      { value: "blue", label: "Blue", checked: false, color: "#0000FF" },
-      { value: "green", label: "Green", checked: false, color: "#008000" },
-      { value: "orange", label: "Orange", checked: false, color: "#FFA500" },
-      { value: "red", label: "Red", checked: false, color: "#C63333" },
-      { value: "purple", label: "Purple", checked: false, color: "#6E20E1" },
-      { value: "yellow", label: "Yellow", checked: false, color: "#FFC300" },
-      { value: "pink", label: "Pink", checked: false, color: "#F463EC" },
-      { value: "brown", label: "Brown", checked: false, color: "#814427" },
-      { value: "grey", label: "Grey", checked: false, color: "#727272" },
-      { value: "black", label: "Black", checked: false, color: "#000" },
-    ],
-  },
-  {
-    id: "gender",
-    name: "Gender",
-    options: [
-      { value: "women", label: "Women", checked: false },
-      { value: "men", label: "Men", checked: false },
-      { value: "kids", label: "Kids", checked: false },
-      { value: "little kids", label: "Little Kids", checked: false },
-    ],
-  },
-  {
-    id: "size",
-    name: "Size",
-    options: [
-      { value: "s", label: "S", checked: false },
-      { value: "m", label: "M", checked: false },
-      { value: "l", label: "L", checked: false },
-      { value: "xl", label: "XL", checked: false },
-      { value: "2xl", label: "2XL", checked: false },
-    ],
-  },
-  {
-    id: "shop by price",
-    name: "Shop by Price",
-    options: [
-      { value: "$0 - $25", label: "$0 - $25", checked: false },
-      { value: "$25 - $50", label: "$25 - $50", checked: false },
-      { value: "$50 - $100", label: "$50 - $100", checked: false },
-      { value: "$100 - $150", label: "$100 - $150", checked: false },
-      { value: "Over $150", label: "Over $150", checked: false },
-    ],
-  },
-];
-
-const Filter = () => {
+const Filter: React.FC<FilterProps> = ({ filters: initialFilters }) => {
   const [filters, setFilters] = useState(initialFilters);
   const [isMobile, setIsMobile] = useState(window.innerWidth < 768);
   const [isDialogOpen, setIsDialogOpen] = useState(false);
@@ -128,22 +75,18 @@ const Filter = () => {
     setIsDialogOpen(!isDialogOpen);
   };
 
+  const applyFiltersAndCloseDrawer = () => {
+    setIsDialogOpen(false);
+  };
+
   function renderFilters() {
     const colorFilters = filters.find((filter) => filter.id === "color");
     const otherFilters = filters.filter((filter) => filter.id !== "color");
 
     return (
       <>
-        <h1 className="mb-4 text-2xl font-bold text-gray-900">Categories</h1>
-        <div>
-          <ul className="mb-6 list-none pl-1 font-bold">
-            {subCategories.map((category, index) => (
-              <li key={index} className="text-gray-700">
-                {category.name}
-              </li>
-            ))}
-          </ul>
-        </div>
+        <h1 className="mb-4 ml-3.5 text-xl font-bold text-gray-900">Filter</h1>
+
         {colorFilters && (
           <Disclosure as="div" className="border-t border-gray-200 py-4">
             {({ open }) => (
@@ -185,7 +128,7 @@ const Filter = () => {
                                 <svg
                                   className="h-6 w-6 fill-current text-white"
                                   xmlns="http://www.w3.org/2000/svg"
-                                  viewBox="0 0 20 20"
+                                  viewBox="0 2 18 20"
                                 >
                                   <path
                                     fillRule="evenodd"
@@ -199,7 +142,7 @@ const Filter = () => {
                                         ? "black"
                                         : "currentColor"
                                     }
-                                    d="M5.293 8.293a1 1 0 011.414 0L10 11.586l3.293-3.293a1 1 0 111.414 1.414l-4 4a1 1 0 01-1.414 0l-4-4a1 1 0 010-1.414z"
+                                    d="M5 11l2 2 6-6 2 2-8 8-4-4"
                                     clipRule="evenodd"
                                   />
                                 </svg>
@@ -255,7 +198,7 @@ const Filter = () => {
                             handleFilterChange(filter.id, option.value)
                           }
                           id={`${filter.id}-${idx}`}
-                          className="h-4 w-4 rounded border-gray-300 text-blue-600 focus:ring-blue-500 "
+                          className="h-5 w-5 rounded border-gray-300 text-blue-600 focus:ring-blue-500 "
                         />
                         <label
                           htmlFor={`${filter.id}-${idx}`}
@@ -289,39 +232,58 @@ const Filter = () => {
             />
             Filter
           </button>
+          <Transition.Root show={isDialogOpen} as={Fragment}>
+            <Dialog
+              as="div"
+              className="relative z-10"
+              onClose={toggleDialog}
+              style={{ fontFamily: "roboto" }}
+            >
+              <Transition.Child
+                as={Fragment}
+                enter="transition-opacity ease-linear duration-300"
+                enterFrom="opacity-0"
+                enterTo="opacity-100"
+                leave="transition-opacity ease-linear duration-300"
+                leaveFrom="opacity-100"
+                leaveTo="opacity-0"
+              >
+                <div className="fixed inset-0 bg-black bg-opacity-25" />
+              </Transition.Child>
 
-          <Dialog
-            as="div"
-            className="relative z-10 "
-            open={isDialogOpen}
-            onClose={toggleDialog}
-          >
-            <div className="fixed inset-0 bg-black bg-opacity-25" />
-            <div className="fixed inset-0 overflow-y-auto">
-              <div className="flex min-h-full items-center justify-center p-4 text-center">
-                <Dialog.Panel
-                  className=" w-full max-w-md transform overflow-hidden rounded-2xl bg-white p-6 text-left align-middle shadow-xl transition-all"
-                  style={{ fontFamily: "roboto" }}
-                >
-                  {renderFilters()}
-                  <div className="mx-5 flex items-center justify-center gap-12  ">
-                    <button
-                      onClick={resetFilters}
-                      className="mt-4 w-[120px] rounded-lg bg-gray-700 px-4 py-2 text-sm text-white  active:scale-95 "
-                    >
-                      Reset Filters
-                    </button>
-                    <button
-                      onClick={resetFilters}
-                      className="mt-4 w-[120px] rounded-lg border-2 border-gray-900 px-3 py-1  text-gray-700 active:scale-95"
-                    >
-                      Apply Filters
-                    </button>
-                  </div>
-                </Dialog.Panel>
+              <div className="fixed inset-0 overflow-y-auto">
+                <div className="flex min-h-full items-center justify-center p-4 text-center">
+                  <Transition.Child
+                    as={Fragment}
+                    enter="transition ease-in-out duration-300 transform"
+                    enterFrom="translate-y-full"
+                    enterTo="translate-y-0"
+                    leave="transition ease-in-out duration-300 transform"
+                    leaveFrom="translate-y-0"
+                    leaveTo="translate-y-full"
+                  >
+                    <Dialog.Panel className="w-full max-w-md transform overflow-hidden rounded-2xl bg-white p-6 text-left align-middle shadow-xl transition-all">
+                      {renderFilters()}
+                      <div className="mx-5 flex items-center justify-center gap-12">
+                        <button
+                          onClick={resetFilters}
+                          className="mt-4 w-[120px] rounded-lg bg-gray-700 px-4 py-2 text-sm text-white active:scale-95"
+                        >
+                          Reset
+                        </button>
+                        <button
+                          onClick={applyFiltersAndCloseDrawer}
+                          className="mt-4 w-[120px] rounded-lg border-2 border-gray-900 px-3 py-1 text-gray-700 active:scale-95"
+                        >
+                          Apply
+                        </button>
+                      </div>
+                    </Dialog.Panel>
+                  </Transition.Child>
+                </div>
               </div>
-            </div>
-          </Dialog>
+            </Dialog>
+          </Transition.Root>
         </>
       ) : (
         <>
